2024-08-26 fix : 1. productInfo 클래스 - 클라이언트에서 file url을 굳이 만들어 보낼 필요 없으므로 파일만 보냄       2. reconditioned 클래스 - 작업자, 점검자 받아오도록 수정 + details 의 jason 변수명 서버 엔티티와 일치화

+++ WindowsFormsApp1/Reconditioned.cs
@@ -62,7 +62,7 @@ namespace WindowsFormsApp1
         {
             public long id { get; set; }
             public DateTime date { get; set; }
-            public int productNum { get; set; }
+            public long productNum { get; set; }
             public string productName { get; set; }
             public string serialNum { get; set; }
             public string worker { get; set; }
